apps/pipeline/requirements-gpu.txt
fastapi
uvicorn
pydantic-settings
httpx
opencv-python
numpy
minio
python-multipart
redis
onnxruntime-gpu
